Create new converter to support the short and byte type.

+++ sample/src/main/java/io/ermdev/mapfierj/sample/model/Person.java
@@ -3,11 +3,11 @@ package io.ermdev.mapfierj.sample.model;
 public class Person {
 
     public String name;
-    public int age;
+    public short age;
     public int height;
     public int petId;
 
-    public Person(String name, int age, int height) {
+    public Person(String name, Short age, int height) {
         this.name = name;
         this.age = age;
         this.height = height;
